--- src/components/pages/doc/inkeep-trigger/index.js
@@ -1,3 +1,0 @@
-import InkeepTrigger from './inkeep-trigger';
-
-export default InkeepTrigger;
